Factor out relation replicator module and wire up migrator to it
