+++ rbac/management/group/relation_api_dual_write_group_handler.py
@@ -24,13 +24,15 @@ from django.conf import settings
 from management.group.model import Group
 from management.models import Workspace
 from management.principal.model import Principal
-from management.role.model import BindingMapping, Role
-from management.role.relation_api_dual_write_handler import (
+from management.relation_replicator.relation_replicator import (
     DualWriteException,
-    OutboxReplicator,
     RelationReplicator,
     ReplicationEvent,
     ReplicationEventType,
+)
+from management.role.model import BindingMapping, Role
+from management.relation_replicator.outbox_replicator import (
+    OutboxReplicator,
 )
 from migration_tool.models import V2boundresource, V2role, V2rolebinding
