+++ rbac/management/principal/cleaner.py
@@ -25,7 +25,7 @@ import xmltodict
 from django.conf import settings
 from management.principal.model import Principal
 from management.principal.proxy import PrincipalProxy, external_principal_to_user
-from management.role.relation_api_dual_write_handler import OutboxReplicator
+from management.relation_replicator.outbox_replicator import OutboxReplicator
 from management.tenant_service import get_tenant_bootstrap_service
 from management.tenant_service.tenant_service import TenantBootstrapService
 from rest_framework import status
