+++ rbac/management/relation_replicator/noop_replicator.py
@@ -1,0 +1,9 @@
+from management.relation_replicator.relation_replicator import RelationReplicator, ReplicationEvent
+
+
+class NoopReplicator(RelationReplicator):
+    """Noop replicator."""
+
+    def replicate(self, event: ReplicationEvent):
+        """Noop."""
+        pass
