+++ rbac/management/tenant_service/__init__.py
@@ -1,7 +1,7 @@
 """Tenant management business logic."""
 
 from django.conf import settings
-from management.role.relation_api_dual_write_handler import RelationReplicator
+from management.relation_replicator.relation_replicator import RelationReplicator
 from management.tenant_service.tenant_service import TenantBootstrapService
 from management.tenant_service.v1 import V1TenantBootstrapService
 from management.tenant_service.v2 import V2TenantBootstrapService
